[Chore] #42 - 뷰 네이밍 변경
