+++ ZipZipSa/ZipZipSa/Presentation/ContentView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 struct ContentView: View {
     var body: some View {
         NavigationStack {
-            NavigationLink(destination: TestView()) {
+            NavigationLink(destination: EssentialInfoView()) {
                 VStack {
                     Image(systemName: "globe")
                         .imageScale(.large)
